Increase limit on attendee list
After spending way too much time playing with pagination in facebook
API, I found out that we can just include a GET param to increase the
limit on # of attendees.

FYI I found that the number is often off by a small number, I suspect
this is a result of some users’ privacy settings.